Changes to systemd journal parser to prevent flaky test #2950
