+++ plaso/parsers/systemd_journal.py
@@ -3,14 +3,14 @@
 
 from __future__ import unicode_literals
 
-import lz4.block
-
-from dfdatetime import posix_time as dfdatetime_posix_time
+import lzma
 
 try:
-  import lzma
+  import lz4.block as lz4_block
 except ImportError:
-  from backports import lzma
+  lz4_block = None
+
+from dfdatetime import posix_time as dfdatetime_posix_time
 
 from plaso.containers import events
 from plaso.containers import time_events
@@ -123,7 +123,12 @@ class SystemdJournalParser(dtfabric_parser.DtFabricBaseParser):
             'Unable to parse LZ4 uncompressed size at offset: 0x{0:08x} with '
             'error: {1!s}').format(file_offset + 64, exception))
 
-      data = lz4.block.decompress(
+      if not lz4_block:
+        raise errors.ParseError((
+            'Unable to decompress LZ4 compressed data at offset: 0x{0:08x} '
+            'with error: missing LZ4 support').format(file_offset + 68))
+
+      data = lz4_block.decompress(
           data[8:], uncompressed_size=uncompressed_size)
 
     return data
