+++ tests/parsers/presets.py
@@ -8,8 +8,6 @@ import unittest
 
 from plaso.containers import artifacts
 from plaso.parsers import presets
-from plaso.parsers import manager as parsers_manager
-from plaso.filters import parser_filter
 
 from tests import test_lib as shared_test_lib
 
@@ -164,23 +162,6 @@ class ParserPresetsManagerTest(shared_test_lib.BaseTestCase):
 
   # TODO add tests for ReadFromFile
 
-class PresetsDataTest(shared_test_lib.BaseTestCase):
-  """Tests for the default preset files."""
-
-  def testParsersAndPresets(self):
-    """Tests that all parsers/plugins in the default presets are valid."""
-    presets_file_path = self._GetDataFilePath(['presets.yaml'])
-
-    preset_manager = presets.ParserPresetsManager()
-    preset_manager.ReadFromFile(presets_file_path)
-    filter_helper = parser_filter.ParserFilterExpressionHelper()
-
-    for name in preset_manager.GetNames():
-      expanded_preset = filter_helper.ExpandPresets(preset_manager, name)
-      _, invalid_parser_elements = (
-          parsers_manager.ParsersManager.CheckFilterExpression(expanded_preset))
-      self.assertFalse(
-          invalid_parser_elements, msg='Invalid parser/plugin name(s)')
 
 if __name__ == '__main__':
   unittest.main()
